refactor: dedupe mixin TYPE_CHECKING stubs into one _CoverHost surface
The movement, lifecycle, calibration and switch-echo mixins each declared
an in-class 'if TYPE_CHECKING' block listing the CoverTimeBased members they
borrow. 41 of those members were repeated across two to four mixins, so a new
shared dependency had to be added in every block that used it.

Collect the whole borrowed surface (112 members) into one _CoverHost stub in
cover_host.py. Each mixin names it as a base only under TYPE_CHECKING
(_MixinBase = _CoverHost, else object), so the runtime base stays object: no
change to any mixin's __bases__ or to CoverTimeBased's MRO. A guard test locks
that invariant.

Behaviour-preserving: pyright reports the same two pre-existing errors and the
full suite passes. Internal only, no CHANGELOG.
