+++ custom_components/cover_time_based/cover_host.py
@@ -1,0 +1,186 @@
+"""Typing-only host surface shared by the CoverTimeBased mixins.
+
+The movement, lifecycle, calibration and switch-echo mixins each read members of
+``CoverTimeBased`` that they do not define themselves. ``_CoverHost`` declares
+that shared surface once so a type-checker resolves those accesses, instead of
+each mixin repeating its own ``if TYPE_CHECKING`` block. A mixin names it as a
+base only under ``TYPE_CHECKING`` (the runtime base stays ``object``), so nothing
+here is evaluated at runtime or added to any class.
+"""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING, Any
+
+if TYPE_CHECKING:
+    import asyncio
+
+    from homeassistant.core import HomeAssistant
+
+    from .calibration import CalibrationState
+    from .travel_calculator import TravelCalculator
+
+
+class _CoverHost:
+    """Static surface of ``CoverTimeBased`` that its mixins rely on."""
+
+    if TYPE_CHECKING:
+        hass: HomeAssistant
+        _relay_intent: dict[str, bool]
+        _pending_switch: dict[str, int]
+        _pending_switch_timers: dict[str, Any]
+        _pending_switch_deadlines: dict[str, float]
+        _removed: bool
+        _wait_for_relay_feedback: bool
+        _feedback_armed_entity: str | None
+        _feedback_wait_entity: str | None
+        _feedback_wait_future: asyncio.Future | None
+        _startup_delay_task: asyncio.Task[Any] | None
+        _stop_is_a_tap: bool
+        _triggered_externally: bool
+        _own_echoes_after_confirming_on: int
+        _calibration: CalibrationState | None
+        _tilt_open_switch_id: str | None
+        _tilt_close_switch_id: str | None
+        _tilt_stop_switch_id: str | None
+        _self_initiated_movement: bool
+        travel_calc: TravelCalculator
+        supports_tilt: bool
+        tilt_calc: TravelCalculator
+        _tilt_strategy: Any
+        _delay_task: asyncio.Task[Any] | None
+        _last_command: str | None
+        _endpoint_runon_time: float | None
+        _min_movement_time: float | None
+        _movement_epoch: int
+        _moving_tilt: bool
+        _moving_tilt_motor: bool
+        _close_includes_tilt: bool
+        _tilt_restore_target: int | None
+        _travel_startup_delay: float | None
+        _tilt_startup_delay: float | None
+        _travel_time_close: float | None
+        _travel_time_open: float | None
+        _tilting_time_close: float | None
+        _tilting_time_open: float | None
+        _open_switch_entity_id: str | None
+        _close_switch_entity_id: str | None
+        _missing_entities_label: str
+        _SWITCH_TARGET_ATTRS: tuple[str, ...]
+        entity_id: str
+        _recalibration_epoch: int | None
+        _tilt_restore_active: bool
+        _pending_travel_target: int | None
+        _pending_travel_command: str | None
+        _pending_tilt_target: int | None
+        _pending_tilt_command: str | None
+        _pending_recalibrated_target: int | None
+        _pending_recalibrated_axis: str | None
+        _tilt_mode_str: str | None
+        current_cover_position: int | None
+        current_cover_tilt_position: int | None
+
+        def _log(self, msg: str, *args: Any) -> None: ...
+        def _is_stale_reappearance(self, old_val: Any, new_val: Any) -> bool: ...
+        def _has_tilt_support(self) -> bool: ...
+        @staticmethod
+        def _entity_unavailable(state: Any) -> bool: ...
+        def async_write_ha_state(self) -> None: ...
+        def _neutralize_tracked_movement(self, *, supersede: bool = True) -> None: ...
+        async def _handle_external_attribute_change(self, event: Any) -> None: ...
+        async def _handle_external_tilt_state_change(
+            self, entity_id: Any, old_val: Any, new_val: Any
+        ) -> None: ...
+        async def _handle_external_state_change(
+            self, entity_id: Any, old_val: Any, new_val: Any
+        ) -> None: ...
+        def stop_auto_updater(self) -> None: ...
+        async def _async_handle_command(self, command: str, *_args: Any) -> None: ...
+        async def _abandon_active_lifecycle(self) -> None: ...
+        def _arm_recalibrated_leg(self, target: Any, axis: Any) -> None: ...
+        def _at_endpoint(self, position: Any) -> bool: ...
+        async def _await_confirmation_before_stop(self) -> None: ...
+        def _begin_movement(
+            self,
+            target: Any,
+            coupled_target: Any,
+            primary_calc: Any,
+            coupled_calc: Any,
+            startup_delay: Any,
+            pre_step_delay: float = 0.0,
+        ) -> None: ...
+        def _cancel_delay_task(self) -> bool: ...
+        def _cancel_startup_delay_task(self) -> bool: ...
+        async def _delayed_stop(self, delay: Any) -> None: ...
+        def _is_direction_change(self, command: Any) -> bool: ...
+        async def _neutralize_parked_move(self) -> None: ...
+        def _recalibration_plan(
+            self, recalibrate: Any, position: Any, *, axis: Any
+        ) -> Any: ...
+        async def _release_displaced_tilt_motor(
+            self, was_tilt_motor_move: Any
+        ) -> None: ...
+        def _self_stops_at_endpoints(self) -> bool: ...
+        async def _send_tilt_close(self) -> None: ...
+        async def _send_tilt_open(self) -> None: ...
+        async def _send_tilt_stop(self) -> None: ...
+        async def _settle_before_reversing(self) -> bool: ...
+        async def _start_recalibration_drive(
+            self, axis: Any, target: Any = 100
+        ) -> bool: ...
+        async def _start_tilt_pre_step(
+            self,
+            tilt_target: Any,
+            travel_target: Any,
+            travel_command: Any,
+            restore_target: Any,
+        ) -> None: ...
+        async def _start_travel_pre_step(
+            self, travel_target: Any, tilt_target: Any, tilt_command: Any
+        ) -> None: ...
+        async def _stop_and_settle_before_recalibration_drive(
+            self, command: Any
+        ) -> bool: ...
+        async def _stop_and_settle_tilt_before_recalibration_drive(
+            self, command: Any, was_tilt_motor_move: Any
+        ) -> bool: ...
+        async def _stop_displaced_movement_for_tilt(
+            self, was_tilt_motor_move: Any
+        ) -> None: ...
+        def _stop_travel_if_traveling(self) -> None: ...
+        async def async_stop_cover(
+            self,
+            *,
+            supersede: bool = True,
+            tilt_axis_reported: bool = False,
+            **kwargs: Any,
+        ) -> None: ...
+        def _has_tilt_motor(self) -> bool: ...
+        def position_reached(
+            self, *, positions: tuple[int | None, int | None] | None = None
+        ) -> bool: ...
+        async def _async_persist_position(self) -> None: ...
+        def _claim_tilt_restore(self) -> int: ...
+        def _clear_multiphase_tilt_state(self) -> None: ...
+        async def _direction_change_delay(self) -> None: ...
+        def _disarm_recalibrated_leg(self) -> None: ...
+        def _movement_target(self, closing: bool) -> str | None: ...
+        def _release_tilt_restore(self) -> None: ...
+        def _require_movement_target_available(self, target: str | None) -> None: ...
+        async def set_position(
+            self, position: Any, *, recalibrate: bool = True
+        ) -> None: ...
+        async def set_tilt_position(
+            self, position: Any, *, recalibrate: bool = True
+        ) -> None: ...
+        def start_auto_updater(self) -> None: ...
+        def _supersede_movement(self) -> None: ...
+        def _tilt_movement_target(self, command: str) -> str | None: ...
+        def _tilt_restore_superseded(self, epoch: int) -> bool: ...
+        def _consume_feedback_arm(self) -> str | None: ...
+        async def _wait_for_relay_echo(
+            self, entity_id: str, timeout: float, *, since: float | None = None
+        ) -> float | None: ...
+        async def _send_stop(self) -> None: ...
+        def _supports_stepped_calibration(self) -> bool: ...
+        def _on_known_position(self, position: int) -> None: ...
